bewise_task/requirements.txt
asgiref==3.5.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
